requirements.txt
numpy
scipy
decorator
mako
appdirs
unittest2
nose-ignore-docstring
